Global changes to perform pixel level gradient computation
Adaptations to code in order to generate the hdf5 data sets and the corresponding output images for latter bench-marking of the image gradients in the image pixel support.

+++ source/color_seg_methods.py
@@ -185,7 +185,7 @@ def optimalK(data, nrefs, maxClusters):
 
 def global_sigma(distances, n_points):
     # return max(distances) / (n_points ** (1/np.float(3)))  # 3 is the dimension data (3d color space)
-    return max(distances) * (n_points ** (1/np.float(6)))  # 3 is the dimension data (3d color space)
+    return max(distances) * (n_points ** (1/np.float(5)))  # 3 is the dimension data (3d color space)
 
 # def global_sigma(rag):
 #     aff_matrix = nx.to_numpy_array(rag, weight='weight')
